+++ History.md
@@ -3,6 +3,7 @@ unreleased
 
  * Accept file system paths in `path` instead of URL-encoded paths
  * Remove `send.etag()`; use `etag` in `options`
+ * Remove `send.hidden()`; use `hidden` in `options`
 
 0.6.0 / 2014-07-11
 ==================